Fixes and updates to finished version (0.9)

+++ bin/unin.sh
@@ -1,5 +1,14 @@
-#!/data/local/tmp/recovery/busybox sh
-BUSYBOX=/data/local/tmp/recovery/busybox
+#!/system/bin/sh
+
+# Initialize Busybox path 
+if [ -x "/system/xbin/busybox" ]; then
+  BUSYBOX="/system/xbin/busybox"
+elif [ -x "/system/bin/busybox" ]; then
+  BUSYBOX="/system/bin/busybox"
+else
+  BUSYBOX="/data/local/tmp/recovery/busybox"
+fi
+
 ${BUSYBOX} mount -o remount,rw /system
 
 # Remove cwm
@@ -22,14 +31,30 @@ if [ -e /system/bin/recovery.cpio ]; then
     ${BUSYBOX} rm /system/bin/recovery.cpio
 fi
 
-${BUSYBOX} rm /system/bin/recovery.sh
-
-if [ -e /system/bin/chargemon.bin ]; then
-   ${BUSYBOX} mv /system/bin/chargemon.bin /system/bin/chargemon
+# Remove byeselinux lkm
+if [ -f "/system/lib/modules/selinux_mod.ko" ]; then
+    ${BUSYBOX} rm /system/lib/modules/selinux_mod.ko
 fi
 
+# Remove recovery script from system
+if [ -f "/system/bin/recovery.sh" ]; then
+    ${BUSYBOX} rm /system/bin/recovery.sh
+fi
+
+# Restore chargemon & e2fsck binary files
+if [ -e /system/bin/chargemon.bin ]; then   
+   CHARGEMON=`${BUSYBOX} sed -n 1p /system/bin/chargemon`
+   if [ "${CHARGEMON}" = "#!/system/bin/sh" ] || [ "${CHARGEMON}" = "#!/system/xbin/busybox sh" ]; then
+       ${BUSYBOX} mv /system/bin/chargemon.bin /system/bin/chargemon
+   fi   
+fi
+
+# Don't touch if it isn't a shell script
 if [ -e /system/bin/e2fsck.bin ]; then
-   ${BUSYBOX} mv /system/bin/e2fsck.bin /system/bin/e2fsck
+   E2FSCK=`${BUSYBOX} sed -n 1p /system/bin/e2fsck` # double checks if e2fsck is not a binary..
+   if [ "${E2FSCK}" = "#!/system/bin/sh" ] || [ "${E2FSCK}" = "#!/system/xbin/busybox sh" ]; then
+      ${BUSYBOX} mv /system/bin/e2fsck.bin /system/bin/e2fsck
+   fi
 fi
 
 ${BUSYBOX} mount -o remount,ro /system
